Use process.stdout.write instead of console.log and increase batch size
`console.log` does some extra work to construct the output, and because
it's already a string, we can save some memory by using
`process.stdout.write`. See this issue for more information:
https://github.com/nodejs/node/issues/1741#issuecomment-103827443

This is an attempt to remedy the OOM issues when rendering all the
issues. Increasing the batch size will also help by reducing the number
of GCs.