Default to using babel for parsing
The version of esprima/espree that eslint defaults to using for parsing
still doesn't work super well for ES6. (Details at
http://eslint.org/blog/2014/12/espree-esprima/).

This changes options to default to using babel for parsing. Also
updating the babel related packages here as well.

If any project does not want babel, it can still be overridden in
`.eslintrc` with a `parser` setting.