Keep paths relative to /code
We saw a bug with `.eslintignore`: when we tested whether a file was
being ignored, we passed the relative directory (which is correct, and
produces expected behavior), but when we actually passed the file for
analysis, we were passing the absolute path. Thus, a file that actually
should not have been ignored was detected as ignored by eslint when
asked to analyze it.

The changes here keep file paths relative for the most part: we chdir to
/code before we do just about anything else, so this works well. I also
extracted all the remaining usage of "/code" to a global var just for
organizational purposes.