Remove intra-issue null characters
We have an issue reported by a user trying out the eslint-3 channel that
they're getting an unparseable issue error. From the stderr output, it
looks like the unparseable issue is invalid JSON that ends right in the
middle of some content explaining the issue. It looks like we're
splitting on the null character and accidentally cutting an issue in
half. Looking into the `node_modules` folder, there are a few places
where we reference the null character and I think it's possible that
we're inserting it in some places that we don't want to.