Update fingerprint for complexity issues targeting anonymous functions
As of https://github.com/eslint/eslint/pull/8058, the message for
complexity violations targeting anonymous functions adhere to the
following format:

> Function has a complexity of {{complexity}}.

This commit updates the fingerprint generated for complexity issues to
address this variation. Unfortunately, because the function is
anonymous, a name can't be included within the fingerprint calculation.

This means that any anonymous function within the same file will receive
the same fingerprint and if that fingerprint is ignored, all complexity
issues for anonymous functions in the file will be ignored.